% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check_cmdstan_installed}
\alias{check_cmdstan_installed}
\title{Check if CmdStan is Installed}
\usage{
check_cmdstan_installed()
}
\value{
Logical, TRUE if installed
}
\description{
Check if CmdStan is Installed
}
\keyword{internal}
