% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonization.R
\name{extract_bmm_data}
\alias{extract_bmm_data}
\title{Extract Data for BMM Analysis}
\usage{
extract_bmm_data(harmonised_dat)
}
\arguments{
\item{harmonised_dat}{Output from harmonize_for_mr()}
}
\value{
List with beta_X, beta_Y, se_X, se_Y, snps
}
\description{
Extract cleaned data from harmonized output for BMM
}
