% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_theta_estimates}
\alias{get_theta_estimates}
\title{Extract theta Estimates}
\usage{
get_theta_estimates(object)
}
\arguments{
\item{object}{bmm_result object}
}
\value{
Data frame with theta estimates from both models
}
\description{
Extract theta Estimates
}
