% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.bmm_result}
\alias{print.bmm_result}
\title{Print Method for BMM Results}
\usage{
\method{print}{bmm_result}(x, ...)
}
\arguments{
\item{x}{bmm_result object}

\item{...}{Additional arguments passed to summary}
}
\value{
Invisibly returns x
}
\description{
Print Method for BMM Results
}
