% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmm_core.R
\name{weighted_mode_estimate}
\alias{weighted_mode_estimate}
\title{Weighted mode estimate}
\usage{
weighted_mode_estimate(
  beta_X,
  beta_Y,
  se_X,
  se_Y,
  phi = 1,
  assume_NOME = FALSE,
  verbose = TRUE
)
}
\description{
Internal weighted mode estimator used in BMM.
}
\keyword{internal}
